RleEncode

+++ oneflow/core/kernel/rle_util.h
@@ -6,7 +6,7 @@
 
 namespace oneflow {
 
-size_t RleEncode(uint8_t* buf, const uint8_t* mask, size_t h, size_t w);
+size_t RleEncode(uint32_t* buf, const uint8_t* mask, size_t h, size_t w);
 }
 
 #endif  // ONEFLOW_CORE_KERNEL_RLE_H_
